backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.2.0
langchain-cohere>=0.3.0
langchain-qdrant>=0.2.0
langgraph>=0.2.0
qdrant-client>=1.11.0
pymupdf>=1.24.0
ragas>=0.2.0
tavily-python>=0.5.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0

